requirements.txt
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
Markdown==3.5.2
psycopg2==2.8.6
pytest==8.0.0
